requirements.txt
python-telegram-bot==20.7
google-cloud-aiplatform
vertexai
python-dotenv
